function/project-python3/requirements.txt
google-cloud-firestore==2.7.3
jsonpickle==3.0.1
